server/api/requirements.txt
fastapi
uvicorn
sentence_transformers
faiss-cpu
google-generativeai